pyproject.toml
[tool.poetry]
name = "talkbut"
version = "0.1.0"
description = "TalkBut - Automated daily work logger and reporter for developers"
authors = ["TalkBut Team"]
readme = "README.md"
packages = [{include = "talkbut", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.0"
gitpython = "^3.1.0"
google-generativeai = "^0.3.0"
pyyaml = "^6.0"
python-dateutil = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.7.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
talkbut = "talkbut.cli.main:cli"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
